requirements.txt
Django==5.1.6
djangorestframework==3.15.2